+++ spec/support/test_scrapers/twitter_scraper.rb
@@ -3,7 +3,7 @@ require 'sagrone_scraper/base'
 class TwitterScraper < SagroneScraper::Base
   TWITTER_PROFILE_URL = /^https?:\/\/twitter.com\/(\w)+\/?$/i
 
-  def self.can_parse?(url)
+  def self.can_scrape?(url)
     url.match(TWITTER_PROFILE_URL)
   end